*** empty log message ***

+++ web_module/branches/branch_hs_2002_11_25/src/main/com/topcoder/web/hs/controller/requests/Base.java
@@ -35,7 +35,7 @@ public abstract class Base implements RequestProcessor {
 
     protected void buildSessionInfo() {
         SessionInfoBean si = new SessionInfoBean();
-        si.setUserId(user.getId());
+        si.setUserId((int)user.getId());
         si.setHandle(user.getUserName());
         si.setGroup(user.getId()==-1 ? 'G' : 'S');  //@@@
         si.setRating(2500);  //@@@
